# Update performance of git repository