Faster look up time for **Last commit date** and **ahead & behind**

**GetLastCommitDate** from 42ms to >0ms
**GetLocalAheadBehind** from 40ms to >2ms